[ZEPPELIN-3075]Fix unqiue algo for the web side in pivot.js file.
### What is this PR for?
unique() algorithm is not correct in pivot.js file.
If the input is `[2, 3, 3, 3, 4, 5]` and the output will be `[2, 3, 3, 4, 5]`. The number `3` is still duplicated.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3075   [ZEPPELIN-3075]

### How should this be tested?
This is very easy and no need test.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: liguohui <liguohui@cmss.chinamobile.com>

Closes #2685 from liguohuicmss/pivot-unqiue-algo and squashes the following commits:

2063175d5 [liguohui] delete the multi empty line
ea582d95e [liguohui] delete some spaces at the end of the line
79c763a6b [liguohui] add a empty line
99cf93da9 [liguohui] Revert "Chart field is also draggable and sortable in the 'keys', 'groups' and 'values'"
fdde39f52 [liguohui] add unit test for unique algo in pivot.js
f99674724 [liguohui] Chart field is also draggable and sortable in the 'keys', 'groups' and 'values'
943e80a96 [liguohui] Fix unqiue algo for the web side in pivot.js file.